Arreglar CPE

+++ src/main/java/com/github/steveice10/mc/classic/protocol/packet/client/ClientCustomBlockSupportLevelPacket.java
@@ -28,7 +28,7 @@ public class ClientCustomBlockSupportLevelPacket implements Packet {
 
     @Override
     public void write(NetOutput out) throws IOException {
-        out.writeByte(this.supportLevel);
+        out.writeByte((byte) this.supportLevel);
     }
 
     @Override
